# docs/contracts/api-contract.yaml
openapi: "3.0.3"
info:
  title: "Lead Management System (NBFC) API"
  version: "1.0.0"
  description: >
    REST API. Uniform envelope {data, meta, error}. Error codes ONLY from docs/contracts/error-taxonomy.md
    (BRD §8.4; VALIDATION_ERROR=400). AuthN: JwtAuthGuard (global); AuthZ: AbacGuard + EntitlementService
    (docs/contracts/auth-matrix.json). Pagination: page (default 1) + limit (default 25, max 100).
    State-creating POSTs accept an Idempotency-Key header. Every endpoint tags the FRs it serves (x-frs).
servers:
  - url: /api/v1
security:
  - bearerAuth: []   # global default; @Public() endpoints override with security: []

# ---- FR coverage map (every FR has >=1 entry; cross-cutting/internal FRs noted) ----
...
paths:
...
  /admin/break-glass:
    get: { summary: "List break-glass grants (DPO/ADMIN)", operationId: listBreakGlass, tags: [Admin], x-frs: [FR-003], parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { name: status, in: query, required: false, schema: { type: string, enum: [pending, active, expired, revoked] } }], responses: { "200": { $ref: '#/components/responses/Ok' }, "403": { $ref: '#/components/responses/Forbidden' } } }
    post: { summary: "Request break-glass grant (DPO/ADMIN)", operationId: breakGlassRequest, tags: [Admin], x-frs: [FR-003], responses: { "201": { $ref: '#/components/responses/Created' }, "400": { $ref: '#/components/responses/ValidationError' }, "403": { $ref: '#/components/responses/Forbidden' } } }
  /admin/break-glass/{id}/approve:
    post: { summary: "Approve break-glass grant (four-eyes)", operationId: breakGlassApprove, tags: [Admin], x-frs: [FR-003], parameters: [{ $ref: '#/components/parameters/PathId' }], responses: { "200": { $ref: '#/components/responses/Ok' }, "403": { $ref: '#/components/responses/Forbidden' } } }
  /admin/break-glass/{id}/revoke:
    post: { summary: "Revoke a break-glass grant early", operationId: breakGlassRevoke, tags: [Admin], x-frs: [FR-003], parameters: [{ $ref: '#/components/parameters/PathId' }], responses: { "200": { $ref: '#/components/responses/Ok' }, "403": { $ref: '#/components/responses/Forbidden' }, "404": { $ref: '#/components/responses/NotFound' } } }

  # ---------- M2/M3 Capture, Workspace, Identity resolution ----------
...
  /admin/config:
    get: { summary: "List pending config versions (maker-checker queue)", operationId: listPendingConfig, tags: [Admin], x-frs: [FR-132], parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { name: config_type, in: query, required: false, schema: { type: string } }], responses: { "200": { $ref: '#/components/responses/Ok' }, "403": { $ref: '#/components/responses/Forbidden' } } }
  /admin/config/{id}/approve:
    post: { summary: "Approve config change (maker-checker)", operationId: approveConfig, tags: [Admin], x-frs: [FR-132], parameters: [{ $ref: '#/components/parameters/PathId' }], responses: { "200": { $ref: '#/components/responses/Ok' }, "403": { $ref: '#/components/responses/Forbidden' } } }
  /admin/config/{id}/rollback:
    post: { summary: "Roll back config version", operationId: rollbackConfig, tags: [Admin], x-frs: [FR-132], parameters: [{ $ref: '#/components/parameters/PathId' }], responses: { "200": { $ref: '#/components/responses/Ok' } } }

  # ---------- M15 Integration & Events ----------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token (also accepted via httpOnly cookie). Authorisation by AbacGuard/EntitlementService (FR-002)."
  parameters:
    Page: { name: page, in: query, required: false, schema: { type: integer, default: 1, minimum: 1 } }
    Limit: { name: limit, in: query, required: false, schema: { type: integer, default: 25, maximum: 100, minimum: 1 } }
    PathId: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
  schemas:
    Meta:
      type: object
      properties:
        correlation_id: { type: string }
        pagination:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: "From error-taxonomy.md (§8.4)" }
        message: { type: string }
        retryable: { type: boolean }
        fields:
          type: array
          items: { type: object, properties: { field: { type: string }, issue: { type: string } } }
        detail: { type: object, additionalProperties: true }
    ErrorEnvelope:
      type: object
      required: [data, error]
      properties:
        data: { type: object, nullable: true, example: null }
        meta: { $ref: '#/components/schemas/Meta' }
        error: { $ref: '#/components/schemas/ApiError' }
...
  responses:
    Ok:
      description: "Success"
      content: { application/json: { schema: { type: object, properties: { data: {}, meta: { $ref: '#/components/schemas/Meta' }, error: { type: object, nullable: true } } } } }
...
    Created:
      description: "Created"
      content: { application/json: { schema: { type: object, properties: { data: {}, meta: { $ref: '#/components/schemas/Meta' }, error: { type: object, nullable: true } } } } }
...
    ValidationError:
      description: "400 VALIDATION_ERROR (may carry detail.reason=STAGE_GUARD_FAILED)"
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    Forbidden:
      description: "403 FORBIDDEN (incl. CONSENT_MISSING)"
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: "404 NOT_FOUND"
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...